refactor: move vote services in vote app
