+++ apps/polls_management/services/poll_service.py
@@ -9,8 +9,8 @@ from apps.polls_management.exceptions.poll_not_valid_creation_exception import P
 from apps.polls_management.exceptions.poll_does_not_exist_exception import PollDoesNotExistException
 from apps.polls_management.exceptions.poll_not_yet_voted_exception import PollNotYetVodedException
 from apps.polls_management.models.poll_option_model import PollOptionModel
-from apps.polls_management.services.majority_vote_service import MajorityVoteService
-from apps.polls_management.services.vote_service import VoteService
+from apps.votes_results.services.majority_vote_service import MajorityVoteService
+from apps.votes_results.services.vote_service import VoteService
 from apps.polls_management.models.poll_model import PollModel
 
 class PollService:
