+++ apps/polls_management/views/dummy_view.py
@@ -13,10 +13,10 @@ from apps.polls_management.exceptions.vote_does_not_exixt_exception import VoteD
 from apps.polls_management.models.majority_vote_model import MajorityVoteModel
 from apps.polls_management.models.poll_model import PollModel
 from apps.polls_management.models.poll_option_model import PollOptionModel
-from apps.polls_management.services.majority_vote_service import MajorityVoteService
+from apps.votes_results.services.majority_vote_service import MajorityVoteService
 from apps.polls_management.services.poll_service import PollService
 from apps.polls_management.exceptions.poll_option_unvalid_exception import PollOptionUnvalidException
-from apps.polls_management.services.vote_service import VoteService
+from apps.votes_results.services.vote_service import VoteService
 
 def all_polls(request: HttpRequest):
     """
