+++ apps/votes_results/services/majority_vote_service.py
