+++ apps/votes_results/services/vote_service.py
