+++ documentation/code_documentation/polls/services/majotiry_vote_service.md
@@ -1,2 +1,2 @@
 # Majority vote service
-::: apps.polls_management.services.majority_vote_service
+::: apps.votes_results.services.majority_vote_service
