+++ test/polls/test_vote_service.py
@@ -4,7 +4,7 @@ import pytest
 from assertpy import assert_that
 from apps.polls_management.models.poll_model import PollModel
 from apps.polls_management.models.poll_option_model import PollOptionModel
-from apps.polls_management.services.vote_service import VoteService
+from apps.votes_results.services.vote_service import VoteService
 from apps.polls_management.exceptions.poll_option_unvalid_exception import PollOptionUnvalidException
 from apps.polls_management.exceptions.poll_does_not_exist_exception import PollDoesNotExistException
 from apps.polls_management.exceptions.vote_does_not_exixt_exception import VoteDoesNotExistException
